refactors attribute encoders into library
