+++ caveat/attribute_encoding/__init__.py
@@ -1,0 +1,5 @@
+from caveat.attribute_encoding.base import BaseAttributeEncoder, tokenize
+from caveat.attribute_encoding.onehot import OneHotAttributeEncoder
+from caveat.attribute_encoding.tokenise import TokenAttributeEncoder
+
+library = {"onehot": OneHotAttributeEncoder, "tokens": TokenAttributeEncoder}
