+++ tests/test_attributes_onehot_encode.py
@@ -4,34 +4,7 @@ from pandas.testing import assert_frame_equal
 from torch import Tensor
 from torch.testing import assert_close
 
-from caveat.encoding.attributes import (
-    OneHotAttributeEncoder,
-    onehot_encode,
-    ordinal_encode,
-)
-
-
-def test_encode_ordinal():
-    data = pd.Series([34, 96, 15])
-    min, max = 0, 100
-    encoded = ordinal_encode(data, min, max)
-    expected = Tensor([[0.34], [0.96], [0.15]]).float()
-    assert_close(encoded, expected)
-
-
-def test_encode_nominal_no_encodings():
-    data = pd.Series(["M", "F", "F"])
-    encoded, encodings = onehot_encode(data, None)
-    assert_close(encoded, Tensor([[0, 1], [1, 0], [1, 0]]).float())
-    assert encodings == {"M": 1, "F": 0}
-
-
-def test_encode_nominal_with_encodings():
-    data = pd.Series(["M", "F", "F"])
-    encodings = {"M": 0, "F": 1}
-    encoded, encodings = onehot_encode(data, encodings)
-    assert_close(encoded, Tensor([[1, 0], [0, 1], [0, 1]]).float())
-    assert encodings == {"M": 0, "F": 1}
+from caveat.attribute_encoding.onehot import OneHotAttributeEncoder
 
 
 def test_encoder_ordinal():
